requirements.txt
uvicorn==0.34.0
fastapi==0.115.8
python-keycloak==5.3.1
line-bot-sdk==3.16.0
PyJWT==2.10.1
python-multipart==0.0.20
sqlmodel==0.0.22
